proof-language: MetaVar: drop well formedness criteria from MetaVars
If the user decides to construct an unsatisfiable metavar, we should not stop them.
It will of course be impossible to completely instantiate this later.
Note that we type checking takes care that we only allow MetaVars or Substs

I think we should go further. Should we disallow redundant substitutions completely?

```
class SSubst:
    ...
    def well_formed():
        # We do not allow redundant substitutions, since it reduces need the meta-inference for freshness etc.
        if var == plug:           return False  # This is the identity substitution, and so is redundant.
        if pattern.s_fresh(var):  return False  # The variable does not occur, so this is redundant
        return super.well_formed()
```

We then simplify the code for `positivity`, freshness etc for these classes.