Proof expressions and pretty printing (#61)
This PR implements proof expressions and interpreters to avoid storing
proofs in memory.
It sets up various interpreters for proofs that implement various
functionality such as serialization and pretty printing.

Functionality wise, this does not add much, so we reuse existing tests.
The formatting implementation is completely rewritten. I only have a
system test, though it could use a unit test as well.

---------

Co-authored-by: Adam Fiedler <info@fiedlr.sk>