Sack pretty printing implementation (#75)
Stack pretty printing is aimed at helping with debugging the proofs
(solves #63). The straightforward solution could be just to print string
representation of patterns from the top of the stack. But this is proven
to be completely unreadable by humans. Even simple patterns look too
cumbersome. The second attempt is to give a string representation to
patterns that makes them easy to recursively convert to string formulas.
This is working, but what is missing in the solution is that each theory
like `proposition.py` has its own notations. So the resulting attempt
allows plugging in the notations to pretty print stack using also string
representation of patters. I agree that it is a bit ugly since the
interpreter is a part of a theory not backward, but maybe together we
can improve the architecture.

Explicit things to look at during the PR review:

- [x] Added decorator to reduce the amount of repeated code and prevent
the possibility to interfere with the underlying logic in serializing
interpreter.
- [x] String representations - better to use known notations, which I
might not be aware of yet.
- [x] Plugging notation into the interpreter.