Deserializing  (#106)
Addresses #32 . Adds a deserializing method which applies the
`PrettyPrintingInterpreter` to proofs in byteformat. Some notes:

- Currently the `StatefulInterpreter` does not consume `MetaVar`
arguments from the stack (nor is there a way to push these arguments to
the stack). The deserializer handles this on its own.
- Currently the `PrettyPrintingInterpreter` uses notation in its output.
I've introduced a `Notationless` interpreter which ignores notation and
uses only memory indices. This interpreter's output should then coincide
with `PrettyPrinting` a byte-stream (which forgets notation).
- A further note regarding notation/memory: in several places we only
look at the pattern and look for it in memory, but we don't assert any
correlation between that location and memory and the passed `id`.
Technically, calls with wrong `id`s can still succeed, which can be
confusing at some point. Perhaps we want to tighten these checks.
- I introduced a class of tests which tests serializing, deserializing
and pretty-printing against all claims and proofs in `Propositional`.

---------

Co-authored-by: Nishant Rodrigues <nishantjr@gmail.com>