Make instantiations use static values (#114)
This introduces taking static values as the values for instantiations,
which is expected and is not currently working in the DSL. The gist of
it is that we write:

```python
proved.instantiate(1: EVar(1))`
```

instead of

```python
proved.instantiate(1: self.evar(1))
```

The advantage is that it is much easier to use and the correct order of
evaluation is handled for us, so it's also easy to extract subproofs and
work with proof constructors.

Disadvantage is that now the patterns are in memory and need to be
"parsed" to create the expression, which might cause slightly worse
performance in the generation phase.
Also, notation is currently disabled, as we cannot use `.load` in such
static expressions, which is why the proofs are longer. It would lead to
"double-loading". This will be handled in future PRs, probably by
introducing a special `Notation` subclass of `Pattern`.