Convert definitions of the notation from a Metamath slice to the new format (#135)
PR adds the implementation of converting Metamath parsed notation terms
to the new format as lambda functions returning Patterns. These
functions can be used to produce patterns with required variables
provided as arguments. Closes #124.

Added files and functionality (only Metamath translator files have been
changed)
* converter.py - added tracking of a variable scope and converting Term
statements to the new format. It is not claimed that any pattern can be
translated, but it works for the slice used for testing. Additional
cases can be added later. Also, symbol discovery was changed to detect
symbols registered using axioms.
* scope.py - introduces a new representation of a scope, which currently
is used to switch from the context of global Metavariables to arguments
provided in Notation definitions. It will also be used to capture
essential hypotheses at parsing axiomatic terms.
* vardict.py - auxiliary implementation of a dictionary to simplify
checking that a Metavariable object is saved in the dictionary by using
the object or just its name attribute. It also helps to implement unit
tests.
* test_converter.py - the test for notation translating is added and the
test for symbols discovery is changed. Other tests were refactored
because variables are now collected to the scope instead of the
Converter class.