Add more wellformedness checking to the checker (#137)
Implements checking disjointness of `e_fresh` and `app_ctx_holes` for
well-formedness of metavars, the reason for this is explained in a
separate proof that we will need to go over. Moreover, some sufficient
well-formedness checking for other kinds of patterns is implemented.

Things to focus on during CR:

- [ ] Double check if the well_formedness checking for other patterns
than the "problematic" cases is fine, using an inductive argument on
putting terms on stack (and assuming `Gamma` contains concrete patterns
only).
For example, adding `mu X. phi` only needs to check positivity of `phi`
and not wellformedness of `phi`, as `phi` is either a primitive (like
`EVar`) or another non-primitive (like `mu`) that we checked with the
given `wellformed` function.
- [ ] Consider my suggestion of moving MetaVar closer to atoms also wrt
instruction numbers.

**Notes**
- Holes checking in `instantiate` will be implemented in a future PR
with Framing.