Add Substitutions to generator (#143)
Adds the `Pop`, `ESubst` and `SSubst` instructions in the generator. 

A few notes about applying substitutions:
- Substitutions are applied maximally, only resorting to wrapping a term
`T` into a `ESubst` constructor when the corresponding substitution
cannot be concretely applied to `T`. Note that `T` here is not
necessarily immediately a `MetaVar`, but can be an arbitrary chain of
substitution constructors on top of a `MetaVar`.
- My intention for the checker (and the generation code which simulates
the checker, like interpreters) is that all constructed patterns are in
a `normalized` state, i.e all substitutions have been applied maximally
to them.
- We should be able to get away with not checking explicitly if a
pattern is `normalized`, but rather maintain this property as an
invariant of the checker code.
- Maintaining this invariant implies modifying the behavior of
`Instantiate`. Instantiating a `MetaVar` might mean that substitution
constructors on top of it may now be applied. Assuming again that
`plugs` used in substitutions and instantiations are `normalized`, it
should be correct to simply apply the substitutions (from bottom to top)
after applying the instantiation to the `MetaVar`. The logic here is
seen in the `instantiate` method for substitution constructors.