Add support for axiom essential hypotheses  (#156)
This is the second piece of work related to support of axioms (addresses
#151). The PR adds tests and implementation for converting axioms with
essential hypotheses to the new format. Short description of changes:

representation.py - adds a class for Axioms with antecedents.
scope.py - adds a method for replacing a metavariable with a new one
with different metaconditions.
converter.py - entirely changed the algorithm for parsing axioms. Now it
allows having intermediate scope based on the result of parsing
metaconditions. Another change is that hard-coded handling of some
notations is replaced by built-in notations.
test_converter.py - adds tests for axioms with EH based on
`disjoint-lemma-slice`.

---------

Co-authored-by: Nishant Rodrigues <nishantjr@gmail.com>