Treat some of the built-in notations as symbols (#200)
# Code review

I'm keeping the built-in definitions of:

1. Implication: This is because of rules like ModusPonens, that actually
tests for one of the patterns to be an implication (rather than a symbol
application)
2. Mu: we have to keep wellformedness checks for any patterns
constructed with mu, so they cannot be treated as symbols. Also, KT rule
creates actual mu-binders and so it might happen that it wouldn't match
up with mu's treated as symbols.
3. Exists: The exists quantifier creates an actual exists binder, it
might not match if exists is used in some axiom that is not desugared.
4. Application: Here I'm not 100% sure if it's needed, but I wanted to
avoid some "circular" issues where application would be treated as an
application of an application ....

# TODO in a future PR

Now that I think of it, for this reason we should call MM axioms for
things like `Propagation_OR` etc., so that we stay consistent with
whatever MM does (if or is treated as a symbol).