Checker: Read inputs upfront into memory (#203)
This PR optimizes reading inputs in the checker so that inputs are read
upfront in their entirety before proceeding to verify the goal. The
optimization eliminates the need for stream-processing closures and
having to perform expensive IO operations for each byte in the input
during verification.

One potential caveat of this optimization is the increased memory
requirements, especially when the inputs are very large. The amount of
memory needed now increases linearly with the input size. However, this
should be less of an issue given that inputs are optimized binary files.

The following tables summarize performance measurements (in R0 zkVM
cycles) before and after this optimization.

### Before the Optimization

| | File IO Only | Verification (inc. IO) | Overall | IO to Overall | IO
Cycles per byte |
| ------------- | ------------ | ---------------------- | ------- |
------------- | ------------------ |
| small_theory | 8441 | 59955 | 82324 | 10.3% | 151 |
| propositional | 74266 | 247201 | 269570 | 27.5% | 110 |
| transfer | 734435 | 3194866 | 3221611 | 22.8%. | 106 |


### After the Optimization

| | File IO Only | Verification (inc. IO) | Overall | IO to Overall | IO
Cycles per byte |
| ------------- | ------------ | ------------------------- | --------- |
------------- | ------------------ |
| small_theory | 4818 | 45199 (4818+40381) | 67755 | 7.1% | 86 |
| propositional | 9948 | 165000 (9948+155052) | 187556 | 5.3% | 15 |
| transfer | 42534 | 2228681 (42534+2186147) | 2255613 | 1.9%. | 6. |


### Improvement Highlights

|                          | File IO Only | Overall   |
| ------------- | ------------ | ------- |
| small_theory  | 43%              | 18%     |
| propositional  | 86%              | 30%    |
| transfer           | 94%              | 30%    |