Implement memoization for translated proofs (#208)
The PR adds two interpreters to address #194:

- CountingInterpreter that allows just to make a dry run and collect
statistics about used patterns. It maintains a complexity score for each
pattern using the number of its usage and total number of commands
required to build it on stack, including if some underlying patterns are
already memoized. Then it just uses a greedy stepwise approach to select
as many patterns for memoization as possible to keep the memory size
under 256 elements.
- MemoizingInterpreter overrides the `pattern` function and load
patterns if they are in the memory and memoize if they were selected by
the CountingIntrepreter for memoization and havn't been memoized yet.

The PR doesn't contain tests right now, but it is implicitly tested by
the checker as its application is turned on by default in the converter.