Notations (#229)
1. This first PR implements just the wrapper to improve pretty printing.
Each notation can decide how it is displayed with `__str__`. We can
decide in the future if notations will also have direct access to
memory.
2. Removes dead legacy code for notation in `Propositional` that came
from when `ProofExpr` was first introduced.
3. I reversed the order of the stack back so that we are not confused
when we apply things like `Implication`
4. I add `|-` for Proved instances so that it is clear in pretty stacks
what is `Proved` and what is a `Pattern`
5. Fixes a bug where PrettyPrinter does not realize a notation was
added, by adding `.add_notation()`.
6. Adds `make verify-translated` target back, as there was no
alternative for it in `Makefile` and it was accidentally removed by
@IljaZakharov .
7. Fixes `update-snapshots`.
8. `instantiate_notation` |-> `instantiate_pattern` as it does not have
anything to do with notations

**CR**
- Look at the changed pretty-printed files first, it is much easier to
see what I did.
- This causes the additional imaginary rewrite step `// Notation`. It
does not have to be there, but then we would need to detect the last
step of placing the last notation subterm on stack, and this would be
more complicated.
- If a notation uses another notation, the proof passes without issues,
but this is not detected by the pretty printer for now. We can add it
later.