Commit Gamma and claim to receipt (#284)
Addresses  #280.

# Code review
1) Not very happy with the closure `next_journal_chunk`, but I did not
find a straightforward solution with Rust iterators that does not use
experimental features like `next_chunk`. The thing is that this iterator
would need to know how to change step sizes.
2) No tests yet. We might think of comparing with the respective file
contents, but I would need to extract the functionality into an
interface. Let me know if this is worth doing now.