zkLLVM pattern (#274)
This PR implements the following features:
## **Define and implement `pattern` syntax in C++**
### Instructions
 - [x] Definition
 - [x] Pattern Matching
### Terms
- [x] Pattern Definition  
- [x] Pattern Constructors
- [x] Pattern Functions (`e_fresh`, `s_fresh`, `positive`, `negative`,
`well_formed`)
  - `e_fresh` -> `pattern_e_fresh` ✅ 
  - `s_fresh` ->  `pattern_s_fresh` ✅ 
  - `positive` -> `pattern_positive` ✅ 
  - `negative` -> `pattern_negative` ✅
  - `well_formed ` -> `pattern well_formed `✅
- [x] `Term` and `Entry` definitions
### Notation
- [x] `bot`
- [x] `not` // I've implemented as `negate` due to C++ name mangling
- [x] `forall`  
### Proof Checker: Memory and Stack
- [x] Stack, Claims, and Memory Definition
- [x] Stack Functions (`pop_stack `, `pop_stack_pattern `,
`pop_stack_proved `)

### Testing
- [x] `test_efresh`
- [x] `test_sfresh`
- [x] `test_positivity`
- [x] `test_wellformedness_positive`

## Important notes:
I've tried to make this PR work for the tests, and as it was getting
bigger, I've split the files to make it easier to review them. However,
I admit that this is still a huge PR to review (even though it's mostly
rewriting an already-tested code in another programming language), so
I'm willing to put extra effort into splitting it into a different
(possibly not executable) PR series.