Fix modes and change test interpreter to memo (#301)
1. Changes the test harness to use `memo` even for generated proofs like
`propositional`
2. Fixes a bug that executed all phases at once even when one specified
a phase. Because we share memory between phases, we no longer run them
independently and thus the argument `phase` has been dropped from the
command line.
3. Because we no longer take `phase`, I switched to `output_folder` and
`slice_name` instead of `file_name` so that we can append whatever
format suffix to `slice_name` that is required.
4. I tidied up in `proof.py`. There was a lot of code redundancy. I
added `IOInterpreter` that is a roof for both text and binary formats
outputting to files. As a single interpreter now takes care of all
phases, `IOInterpreter` takes now the output streams for all of them.
The advantage is that `IOInterpreter` switches the current output
automatically when switching phases.