Refactoring zkLLVM Proof Checker (#306)
This simple but important refactoring PR prepares the code for the
`execute_instructions` at #315.

-
[8b1e624](https://github.com/runtimeverification/proof-checker/pull/306/commits/8b1e624fa1d4c8d7aa22c686440a70c0d691bd13):
Removing `LinkedList::Create` with two elements: It looked like a good
idea at first, but it can be confusing where it was created with `push`
or `push_back`
-
[59d941a](https://github.com/runtimeverification/proof-checker/pull/306/commits/59d941a5912d1295a4f9c37d357240ff082eb3ae):
Renaming `insert_front` to push
-
[d8a118e](https://github.com/runtimeverification/proof-checker/pull/306/commits/d8a118e2311de11e47f916c224dc05c8df99bb8e):
Refactoring `delete_front` and `pop`: As `pop` was the only user of
`delete_front`, I choose to inline it there.
-
[d249da5](https://github.com/runtimeverification/proof-checker/pull/306/commits/d249da56ce8a6ab3a75e4a1e087013b52ceb3876):
Refactoring `operator==` of Pattern and LinkedList: Using `==` instead
of `operator==` and deleting DEBUG message
-
[b5d3789](https://github.com/runtimeverification/proof-checker/pull/306/commits/b5d37899446b0a4433c29bd9c1d7b9175f288c39):
Using `int8_t `instead of int for Instructions
-
[8a321a2](https://github.com/runtimeverification/proof-checker/pull/306/commits/8a321a28ebab0129b1be546660705e312aa51904):
Moving destructors and print function to closer of struct boilerplate
-
[26094e4](https://github.com/runtimeverification/proof-checker/pull/306/commits/26094e461050a382f407c5ec50cb10dff2333416):
Deleting `Pattern::` where it ins't necessary
-
[d12f32f](https://github.com/runtimeverification/proof-checker/pull/306/commits/d12f32fa78165a22968210f14d5c188b0643ff4c):
Refactoring `TermType` and `EntryType`: Creating real `Term` and `Entry`
classes with all necessary boilerplate