Reference Counting implementation for zkLLVM (#353)
This PR indents to:
- Simplify the code using only `Term` instead of `Term` and `Entry`
- Simplify the input reading from zkLLVM using the `std::array` from the
arguments of the circuit function as the buffer for the Proof Checker
instead of converting it to a Linked List
- Create a Reference Counting implementation that works as a Shared
Pointer in C++. In this implementation, we called it `Rc` to be similar
to the Rust implementation of the Proof Checker.
- With that, we simplify memory management by not having to free or call
the destructors by the end of the functions, only on the destructors
themselves.
- We also got a simplification on the `operator==` implementations for
free.
- zkLLVM doesn't support the `invoke` opcode, used to call functions
that can throw. To ensure that zkLLVM never uses this opcode, we marked
all functions with the `noexcept` attribute.
- To fix a memory error pointed out several times by `valgrind`, we
implement our own `memset` function (as zkLLVM doesn't support the
library implementation of it) to initialize all fields of the structs
that we alloc memory.
- We're now using `uint8_t` instead of `int` everywhere except on the
array input, as the zkLLVM only reeds `int` values.