Change all propositional proof generation to return proof thunks along with the conclusion of that proof (#362)
This PR addresses a long standing issue regarding the unavailability of
conclusions when working with unevaluated proof generating thunks.

This change now makes all proof builders also return the conclusion of
the proof which upon execution of the thunk is checked against the one
returned by the underlying function (This may prove to be unnecessary
and in the future we may choose to make the thunk just return None,
while keeping the conclusion as it is here, in the unevaluated thunk).

Compared to all other approaches to solve this issue, this is by far the
simplest, safest and the one with the most performance gain.

NOTE THAT this change does not keep the entire proof in memory or
anything like that, it just keeps track of a small pattern representing
the end conclusion of the proof, so there will be no blow-up when
adapting this to larger proofs.