Proofs for Propositional Tautologies via Resolution (Chapter I: Term to CNF) (#233)
This PR is the first in a series of PRs aiming to implement a proof
generator for propositional tautologies. The final function will
effectively implement the Resolution Algorithm (
https://en.wikipedia.org/wiki/Resolution_(logic) ) for our term
representation as well as generate an appropriate proof of the original
pattern (if it is indeed a tautology).
As part of this algorithm we first need to convert the term to an
equivalent one in Conjunctive Normal Form (CNF). This is what this PR
does.