Preliminar version of `execute_instructions` in Lurk (#441)
Part of #416 
This PR introduces the basic implementation for `execute_instructions`,
modifying the necessary functions and implementing new ones.

The idea, for now, is to not enclose `execute_instructions` into a
definition to improve the readability and the development process.

We currently don't have switch-case support in Lurk, so we end up with
the big chained if-statements.

As we don't have variables, only values, we take the arguments, process
them, and return them modified if needed.

The best way to review this is locally on VS Code using the Lurk
extension for code highlighting!