Update hint parser and generated hints (#338)
This PR is a draft that updates the Python parser to accept proof hints
as generated by the current version of K and the LLVM backend (i.e.
including _function events_ as documented in
https://github.com/runtimeverification/llvm-backend/pull/872); it also
regenerates the proof hints stored in the repo to ensure that CI
passes[^1].

I do not intend for this PR to be merged in anything like its current
state. In particular, the following additional changes will need to be
made:
* Testing of the function event fields; currently they are silently
accepted and discarded.
* Possible interleaving and combination of the rewrite steps and
function events; they are both stored as separate sequences at the
moment with no ordering imposed between them.
* Interpretation of the relative position component of the function
events; these are stored only as raw strings at the moment.
* An update (even if it's manual) to the K dependency version to make
sure that everything is in sync.
* Refactoring and cleanup of the Python code introduced here; the
changes made are pretty much the minimal possible set to get myself a
working parser, and we may want to improve the code a bit before
merging.

Some of these changes are blocked on my own understanding of the problem
domain, and I will endeavour to update the parser as I understand things
better. As currently implemented, we ought to be able to use the new
parser to test the changes required in
https://github.com/runtimeverification/llvm-backend/issues/860, as we
were previously unable to do so properly.

### EDIT by @malturki:

The parser is updated to handle the new hints format (without side
conditions). Please note the following:
1. The trace the parser builds does not currently add kore term
arguments to function and hook event objects (places where this should
happen are marked with todos). This will be handled in a future PR
(probably when we add support for side conditions)
2. We added minimal updates to the logic that processes traces generated
by the parser. We currently ignore function and hook events. This is
also marked by todos appropriately.
3. This PR bumps up the K and PyK versions, so it will be important to
ensure the proper versions are installed locally for local testing.


[^1]: Note that the changes to the format in
https://github.com/runtimeverification/llvm-backend/pull/872 are
actually backwards compatible; we could actually parse the old generated
hints with this extended parser.

---------

Co-authored-by: musab.alturki <musab.alturki@gmail.com>
Co-authored-by: Musab Alturki <42160792+malturki@users.noreply.github.com>