Set up K / Pyk automatic dependency updates (#449)
This PR updates the K / Pyk dependency update mechanism for this
repository such that it will behave correctly when the dependency update
job is enabled. The changes are:
* Fixes to the existing dependency update workflow to reflect the
project's current structure.
* A manual update to ~~the latest~~ a more up-to-date version of Pyk / K
to test that the scripts work correctly and that the project is up to
date.
* The newest versions need a change to the K frontend to get merged; I
will take care of this separately.
* ~~Removal of the checked-in binary execution hint files. Because these
correspond to a particular version of K, they should not be checked in
to source control, and instead be generated as part of the integration
testing workflow.~~
* This will need to wait for a separate change to redesign the test
stanzas in the project Makefile; there are places that assume these
files exist.
* ~~In the interim, I have specified that the integration tests should
clean the generated hints out before regenerating and running tests.~~
This doesn't work either; hints need to be regenerated by hand.

Blocked on #338 being merged; @malturki has implemented the required
changes to get the project's usage of Pyk up to date. Once that PR is
merged, this one can be updated and reviewed properly.

---------

Co-authored-by: devops <devops@runtimeverification.com>