Fix metavar interpretation bug (#481)
The way metavars are built in the Basic Interpreter is buggy. The Basic
Interpreter builds patterns for each of its constraint lists but this is
wrong. The Stateful Interpreter for instance would push these
constraints onto the stack but not pop them when building the metavar.
Changing the stateful interpreter to do this would be time consuming (as
it would require a change in architecture) and useless. When dealing
with other constructs that keep track of E/S variable data (such as
Mu/Exists) we already don't build the var as a pattern with `pattern` so
my current change is also keeping in line with this convention which has
never caused us issues so far.