Split Performance Numbers for Lurk (#591)
Splitting GPU time in: Execution +Prove | Verify | Prove + Verify

Unfortunately, afaik is impossible to split between execution time and
prove time in Lurk.
Due to the known issue of not being able to execute Lurk only in the CPU
when GPU is available, we can't update the numbers regarding CPU
execution using the same machine.

---------

Co-authored-by: Yi <yi.zhang@runtimeverification.com>
Co-authored-by: Yi Zhang <zorozy@gmail.com>