Remove sorts from kore-and and kore-or (#616)
Fixes pattern matching over `kore_and` and `kore_or` by removing the
sort argument that is not used in the notation definition, as these do
not contain the sort `phi0` argument in their definition, so pattern
match fails.

See
https://runtimeverification.slack.com/archives/C05MF3KEWNM/p1702398036443069
for more explanation

*Code review*

Please check out this branch and make sure that the sort information was
removed from all the applications of `kore_and` and `kore_or`.

---------

Co-authored-by: Ilja Zakharov <ilja.zakharov@runtimeverification.com>