Sync wellformedness checking with R0 implementation and add wellformedness checking for Notation (#638)
1. Updates the state of well_formedness checking in the DSL with that in
the R0 proof checker using `__post_init__`
2. Adds `occurring_vars()` method on Patterns that keeps track of free
variables actually occurring in the pattern, where metavars have no free
variables unless instantiated with patterns containing them
3. Adds well_formedness checking to `Notation` that asserts the
definition to contain no free variables

*Code review*
- Please think twice why I added the checkings on redundant
ESubst/SSubsts before arguing this complicates things in the DSL. It is
actually making things simple, *exactly* because we can make this
assumption to remove edge cases like the ones I removed from the test
suite. This is not only the case of the `metavars` method, but also the
correctness of `e_fresh` and others depends on the substitutions not
being redundant.

We should add a follow-up task that replaces `ESubst/SSubst`
constructors with `apply_esubst/apply_ssubst` everywhere where it is
applicable.

- I know that it is impossible to compute free variables guarantees
correctly because of metavars, that is why we have freshness
constraints. I'm open to rename `occurring_vars` to whatever as long as
it is doing what I intended: keeping track of actual free occurences
*currently* in the pattern.

I added a note that `fv` makes no guarantees about freshness wrt proving
soundness of the proof system, that's not its purpose.

---------

Co-authored-by: Nishant Rodrigues <nishantjr@gmail.com>